enable installing keylime from local sources

+++ setup/install_upstream_keylime/test.sh
@@ -22,10 +22,16 @@ _EOF'
             RHEL_EXTRA_PKGS="cfssl"
         fi
         rlRun "yum -y install $RHEL_EXTRA_PKGS git-core python3-pip python3-pyyaml python3-tornado python3-simplejson python3-requests python3-sqlalchemy python3-alembic python3-packaging python3-psutil python3-gnupg python3-cryptography libselinux-python3 procps-ng tpm2-abrmd tpm2-tss tpm2-tools python3-zmq patch"
-        rlRun "rm -rf keylime && git clone https://github.com/keylime/keylime.git"
-        pushd keylime
+        if [ -d /var/tmp/keylime_src_dir ]; then
+            rlLogInfo "Installing keylime from /var/tmp/keylime_sources"
+            rlRun "pushd /var/tmp/keylime_sources"
+        else
+            rlLogInfo "Installing keylime from cloned upstream repo"
+            rlRun "rm -rf keylime && git clone https://github.com/keylime/keylime.git"
+            rlRun "pushd keylime"
+        fi
         rlRun "python3 setup.py install"
-        popd
+        rlRun "popd"
     rlPhaseEnd
 
     rlPhaseStartTest "Test installed binaries"
